+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Version 5.1.9 (not yet released)
 
 **Weather sources**
-- [GeoSphere Austria] Added as worldwide main source (not yet available: hourly weather code, alerts)
+- [GeoSphere Austria] Added as main source in Austria and nearby
 - [GeoSphere Austria] Added as secondary source for air quality in Europe and nearby
 - [GeoSphere Austria] Added as secondary source for precipitation nowcasting in Austria and nearby
 - [AccuWeather] Fix fail to refresh when using a language not supported by AccuWeather (noticed on Central Kurdish, Esperanto and Interlingua), now fallbacks to English